Hook refactoring, changed derived source type

+++ src/server/lib/math.js
@@ -5,6 +5,6 @@ module.exports = create(
   {
     matrix: 'Array',
     number: 'BigNumber',
-    precision: 32
+    precision: 15 // Hardcoded!
   }
 )
